add a hack to allow run_pyspelling messages to be parsed by vscode
It looks like VSCode (at least in github codespace) only parses the first
error reported by run_pyspelling.py. Adding a delay at the end seems to
fix this issue, as it might prevent the early exit of whatever is parsing
the command output in VSCode.

See the following issue on github for an idea of the problem:
https://github.com/microsoft/vscode/issues/92868