Fix tasks reporting (#141)
* reformat spelling script

* remove unused variable, use flush option from print(), fix spelling error

* change tone of comment about pyspelling ;-)

* simplify pyspelling postprocessing using flashtext

The flashtext is used to directly find the misspelled words in the
original document, with their column span.

* fix/simplify pattern matcher for spellchecker tasks

- severity doesn't need to be specified as it is deduced from message
- "pattern" doesn't need to be a list
- loop is not needed as each output line is parsed independently
- endColumn need to be specified instead of endLine

* remove unused module

* add a hack to allow run_pyspelling messages to be parsed by vscode

It looks like VSCode (at least in github codespace) only parses the first
error reported by run_pyspelling.py. Adding a delay at the end seems to
fix this issue, as it might prevent the early exit of whatever is parsing
the command output in VSCode.

See the following issue on github for an idea of the problem:
https://github.com/microsoft/vscode/issues/92868

* simplify tasks definition

- no need for loop in problemMatcher
- no need for a list of regex in problemMatcher
- no need for default severity as it is parsed from messages

* use the ugly hack to fix message parsing in vscode on run_meta_check.py

* refactor run_proselint.py (remove dead code and reformat)

* load content file using pathlib in run_proselint.py

As the file is fully loaded anyway, this avoids keeping the file open
more time than necessary.