Update my-nesi-org-nz_release_notes_v2-29-0.md
Condensed some of the explanations. Removed the "(new)" from the storage types as it looked like they were new to the system (rather than new to the form). Added a link to documentation on NeSI storage types. May also be useful to add a link to the last bullet for more info on project member roles (see question posed in #support in Slack).

Signed-off-by: Jana Makar <36392208+janamakar@users.noreply.github.com>