Update MobaXterm moving files
Long story short, the SCP method in MobaXterm if you are using the non-interactive method in MobaXterm. You can use this if you set up mahuika using the interactive procedure, but it is a very cluncy method. It would be better to use WinSCP. Remove this completely with the web-activated method. 

Signed-off-by: Geoffrey Weal <geoffrey.weal@gmail.com>