Update MobaXterm moving files (#820)
Long story short, the SCP plugin in MobaXterm does not work/show if you
are using the non-interactive method in MobaXterm. You can use this if
you set up mahuika using the interactive procedure, but it is a very
clunky method. It would be better to use WinSCP. Remove this completely
with the web-activated method.

---------

Signed-off-by: Geoffrey Weal <geoffrey.weal@gmail.com>
Signed-off-by: Cal <35017184+CallumWalley@users.noreply.github.com>
Co-authored-by: Cal <35017184+CallumWalley@users.noreply.github.com>